Refactor utils into packages
