+++ sheska/src/main/kotlin/ar/pelotude/ohhsugoi/util/image/ImageExceptions.kt
@@ -1,0 +1,12 @@
+package ar.pelotude.ohhsugoi.util.image
+
+enum class DownloadErrorType {
+    DIMENSIONS_EXCEEDED,
+    UNSUPPORTED_FORMAT,
+}
+
+class UnsupportedDownloadException(
+    message: String?,
+    cause: Throwable? = null,
+    code: DownloadErrorType,
+) : Exception(message, cause)
